Expose 'prefetchN' parameter in 'ServerOptions'. Move 'f.g.{client => shared}.StreamIngest'.
